# artifacts/api-server/src/routes/enquiries.ts
import { Router, type IRouter } from "express";
import { desc, eq } from "drizzle-orm";
import { db, enquiriesTable, attractionsTable, locationsTable, countriesTable } from "@workspace/db";
import {
  CreateEnquiryBody,
  CreateEnquiryResponse,
  ListEnquiriesResponse,
  UpdateEnquiryStatusBody,
  UpdateEnquiryStatusParams,
  UpdateEnquiryStatusResponse,
  DeleteEnquiryParams,
  ListEnquiryNotificationsParams,
  ListEnquiryNotificationsResponse,
  ResendNotificationParams,
  SendTestEmailBody,
  SendTestEmailResponse,
} from "@workspace/api-zod";
import { requireAdmin } from "../middleware/requireAdmin";
import { serialize } from "../lib/serialize";
import {
  queueEnquiryNotifications,
  notificationsForEnquiry,
  resendNotification,
  queueSummary,
} from "../lib/notifications";
import { isMailConfigured, mailConfigError, mailStatus, sendMail } from "../lib/mailer";
import { testMessage } from "../lib/templates";
import { toE164, whatsAppStatus } from "../lib/whatsapp";

const router: IRouter = Router();

function optionalText(value: string | null | undefined): string | null {
  const normalized = value?.trim();
  return normalized ? normalized : null;
}

router.post("/enquiries", async (req, res): Promise<void> => {
  const parsed = CreateEnquiryBody.safeParse(req.body);
  if (!parsed.success) {
    req.log.warn({ errors: parsed.error.message }, "Invalid enquiry submission");
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const data = parsed.data;
  const firstName = data.firstName.trim();
  const lastName = data.lastName.trim();
  const email = data.email.trim();
  const notes = optionalText(data.notes);
  const phone = optionalText(data.phone);
  const tourTitle = optionalText(data.tourTitle);
  const tourLocation = optionalText(data.tourLocation);
  const enquiryType = optionalText(data.enquiryType);
  const budget = optionalText(data.budget);

  if (!firstName || !lastName || !email) {
    res.status(400).json({ error: "Name and email are required" });
    return;
  }

  if (data.source === "tour") {
    if (!phone || !data.acceptPrivacy) {
      res.status(400).json({
        error: "A phone number and privacy acceptance are required for tour enquiries",
      });
      return;
    }

    if (
      !tourTitle ||
      !tourLocation ||
      !data.tourDurationDays ||
      !Number.isInteger(data.tourDurationDays)
    ) {
      res.status(400).json({
        error: "Tour title, location, and a whole-day duration are required for tour enquiries",
      });
      return;
    }

    if (enquiryType || budget) {
      res.status(400).json({
        error: "Tour enquiries cannot include contact-only context",
      });
      return;
    }
  }

  /*
   * An attraction enquiry names the attraction by id, and the name and place
   * stored with it are read from the database, not taken from the form. They
   * are what the office sees and what the emails print, so they should say
   * what the attraction is called, not what the browser sent.
   */
  let attractionId: number | null = null;
  let attractionTitle: string | null = null;
  let attractionPlace: string | null = null;
  if (data.source === "attraction") {
    if (!phone || !data.acceptPrivacy) {
      res.status(400).json({
        error: "A phone number and privacy acceptance are required for attraction enquiries",
      });
      return;
    }
    if (!data.attractionId) {
      res.status(400).json({ error: "Which attraction is this enquiry about?" });
      return;
    }
    if (enquiryType || budget || data.tourDurationDays) {
      res.status(400).json({ error: "Attraction enquiries cannot include tour or contact-only context" });
      return;
    }
    const [a] = await db
      .select({
        id: attractionsTable.id,
        name: attractionsTable.name,
        location: locationsTable.name,
        country: countriesTable.name,
      })
      .from(attractionsTable)
      .innerJoin(locationsTable, eq(locationsTable.id, attractionsTable.locationId))
      .leftJoin(countriesTable, eq(countriesTable.id, locationsTable.countryId))
      .where(eq(attractionsTable.id, data.attractionId));
    if (!a) {
      res.status(400).json({ error: "That attraction is no longer available" });
      return;
    }
    attractionId = a.id;
    attractionTitle = a.name;
    attractionPlace = [a.location, a.country].filter((v, i, arr) => v && arr.indexOf(v) === i).join(", ");
  } else if (data.attractionId) {
    res.status(400).json({ error: "Only attraction enquiries can name an attraction" });
    return;
  }

  if (data.source === "contact") {
    if (!notes || notes.length < 10) {
      res.status(400).json({
        error: "Please provide a message of at least 10 characters",
      });
      return;
    }

    if (!enquiryType) {
      res.status(400).json({ error: "An enquiry type is required" });
      return;
    }

    if (
      tourTitle ||
      tourLocation ||
      data.tourDurationDays ||
      data.title ||
      data.isTravelAdvisor !== undefined
    ) {
      res.status(400).json({
        error: "Contact enquiries cannot include tour-specific context",
      });
      return;
    }
  }

  const [enquiry] = await db
    .insert(enquiriesTable)
    .values({
      source: data.source,
      title: optionalText(data.title),
      firstName,
      lastName,
      email,
      phone,
      isTravelAdvisor: data.isTravelAdvisor ?? null,
      notes,
      acceptPrivacy: data.acceptPrivacy,
      receiveUpdates: data.receiveUpdates,
      tourTitle: attractionTitle ?? tourTitle,
      tourLocation: attractionPlace ?? tourLocation,
      tourDurationDays: data.tourDurationDays ?? null,
      attractionId,
      enquiryType,
      budget,
      whatsappConsent: data.whatsappConsent ?? false,
      /*
       * Normalised once, here, rather than at send time. A number we cannot
       * make sense of is then visibly null on the enquiry — which is the whole
       * signal, because sending a stranger's travel plans to a mistyped number
       * is worse than not sending at all.
       */
      phoneE164: toE164(phone),
    })
    .returning();

  /*
   * Queue the confirmation and the office alert, but do not wait for them.
   * Sending happens in the background precisely so that a slow or unreachable
   * mail server cannot turn a saved enquiry into an error for the visitor.
   */
  void queueEnquiryNotifications(enquiry);

  res.status(201).json(CreateEnquiryResponse.parse(serialize(enquiry)));
});

router.get("/admin/enquiries", requireAdmin, async (_req, res): Promise<void> => {
  const enquiries = await db
    .select()
    .from(enquiriesTable)
    .orderBy(desc(enquiriesTable.createdAt));

  res.json(ListEnquiriesResponse.parse(serialize(enquiries)));
});

router.patch(
  "/admin/enquiries/:id/status",
  requireAdmin,
  async (req, res): Promise<void> => {
    const params = UpdateEnquiryStatusParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }

    const parsed = UpdateEnquiryStatusBody.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    const [enquiry] = await db
      .update(enquiriesTable)
      .set({
        status: parsed.data.status,
        handledAt: parsed.data.status === "handled" ? new Date() : null,
      })
      .where(eq(enquiriesTable.id, params.data.id))
      .returning();

    if (!enquiry) {
      res.status(404).json({ error: "Enquiry not found" });
      return;
    }

    res.json(UpdateEnquiryStatusResponse.parse(serialize(enquiry)));
  },
);

router.delete(
  "/admin/enquiries/:id",
  requireAdmin,
  async (req, res): Promise<void> => {
    const params = DeleteEnquiryParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }

    /*
     * The enquiry's notification rows go with it, by the cascade on
     * notifications.enquiry_id. They record messages sent about this enquiry
     * and describe nothing once it is gone.
     */
    const [row] = await db
      .delete(enquiriesTable)
      .where(eq(enquiriesTable.id, params.data.id))
      .returning();

    if (!row) {
      res.status(404).json({ error: "Enquiry not found" });
      return;
    }
    res.sendStatus(204);
  },
);

/* ==================================================================== *
 * Notification delivery
 * ==================================================================== */

router.get(
  "/admin/enquiries/:id/notifications",
  requireAdmin,
  async (req, res): Promise<void> => {
    const params = ListEnquiryNotificationsParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }
    const rows = await notificationsForEnquiry(params.data.id);
    res.json(ListEnquiryNotificationsResponse.parse(serialize(rows)));
  },
);

router.post(
  "/admin/notifications/:id/resend",
  requireAdmin,
  async (req, res): Promise<void> => {
    const params = ResendNotificationParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }
    const ok = await resendNotification(params.data.id);
    if (!ok) {
      res.status(404).json({ error: "Notification not found" });
      return;
    }
    // 202: accepted for sending, not sent yet — the worker owns that.
    res.sendStatus(202);
  },
);

/**
 * Which channels will actually send, and what is waiting.
 *
 * Exists because "is it configured" and "does it work" are different questions,
 * and the honest answer to the second one lives in a boot log nobody reads. A
 * wrong password is configuration that looks complete and sends nothing; this
 * reports what the provider said when the server asked it.
 */
router.get(
  "/admin/notification-channels",
  requireAdmin,
  async (_req, res): Promise<void> => {
    const queue = await queueSummary();
    res.json({ channels: [mailStatus(), whatsAppStatus()], ...queue });
  },
);

router.post(
  "/admin/notifications/test",
  requireAdmin,
  async (req, res): Promise<void> => {
    const parsed = SendTestEmailBody.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    if (!isMailConfigured()) {
      /*
       * 503 rather than 500: nothing is broken, the server simply has no mail
       * credentials. The message names which ones are missing so this is
       * actionable without reading the server log.
       */
      res.status(503).json({
        error: mailConfigError() ?? "Email is not configured on this server",
      });
      return;
    }

    /*
     * Sent inline rather than queued, because the entire point is to find out
     * now whether the credentials work. A queued test would report success the
     * moment it was written to the table, which proves nothing.
     */
    try {
      // The same template the admin screen previews, so a test proves the
      // branded shell renders too — not just that the credentials work.
      const message = await testMessage();
      const result = await sendMail({
        to: parsed.data.to.trim(),
        subject: message.subject,
        text: message.body,
        html: message.html,
      });
      res.json(SendTestEmailResponse.parse(result));
    } catch (err) {
      res.status(503).json({
        error: err instanceof Error ? err.message : "Sending failed",
      });
    }
  },
);

export default router;