# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MNT Embark API — luxury tour company
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enquiries:
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a customer enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "201":
          description: Enquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "400":
          description: Invalid enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/enquiries/{id}/notifications:
    get:
      operationId: listEnquiryNotifications
      tags: [enquiries]
      summary: Messages queued or sent for an enquiry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /admin/notifications/{id}/resend:
    post:
      operationId: resendNotification
      tags: [enquiries]
      summary: Put a message back in the send queue
      description: >
        Resets the attempt counter - a resend follows a human fixing something,
        so it gets a fresh set of attempts rather than inheriting an exhausted
        one.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "202":
          description: Queued for sending
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/notification-channels:
    get:
      operationId: getNotificationChannels
      tags: [enquiries]
      summary: Which channels will send, and what is waiting
      security:
        - adminSession: []
      responses:
        "200":
          description: One entry per channel, plus the queue counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannels"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/notifications/test:
    post:
      operationId: sendTestEmail
      tags: [enquiries]
      summary: Send a test email to prove SMTP works
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestEmailInput"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestEmailResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Email is not configured on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List customer enquiries for admins
      security:
        - adminSession: []
      responses:
        "200":
          description: Enquiries ordered newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/enquiries/{id}:
    delete:
      operationId: deleteEnquiry
      tags: [enquiries]
      summary: Delete an enquiry and its notification history
      description: >
        Permanent. The notification rows go with it — they describe messages
        sent about this enquiry and mean nothing without it.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/enquiries/{id}/status:
    patch:
      operationId: updateEnquiryStatus
      tags: [enquiries]
      summary: Update an enquiry lifecycle status
      security:
        - adminSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryStatusInput"
      responses:
        "200":
          description: Updated enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Enquiry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Stats ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Itinerary Step ─────────────────────────────────────────────────
...
    ChannelStatus:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        configured:
          type: boolean
        ready:
          type: boolean
        detail:
          type: ["string", "null"]
        identity:
          type: ["string", "null"]
        checkedAt:
          type: ["string", "null"]
      required: [key, label, configured, ready, detail, identity, checkedAt]
    NotificationChannels:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: "#/components/schemas/ChannelStatus"
        queued:
          type: number
        failed:
          type: number
      required: [channels, queued, failed]
    Notification:
      type: object
      properties:
        id:
          type: number
        enquiryId:
          type: ["number", "null"]
        channel:
          type: string
        templateKey:
          type: string
        recipient:
          type: string
        subject:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, sent, failed]
        attempts:
          type: number
        lastError:
          type: ["string", "null"]
        body:
          type: string
        bodyHtml:
          type: ["string", "null"]
        payload:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
        sentAt:
          type: ["string", "null"]
      required:
        [
          id,
          enquiryId,
          channel,
          templateKey,
          recipient,
          subject,
          body,
          status,
          attempts,
          lastError,
          createdAt,
          sentAt,
        ]
    TestEmailInput:
      type: object
      properties:
        to:
          type: string
          minLength: 3
      required: [to]
    TestEmailResult:
      type: object
      properties:
        messageId:
          type: string
      required: [messageId]
    Enquiry:
      type: object
      properties:
        id:
          type: number
        source:
          type: string
          enum: [tour, contact, attraction]
        status:
          type: string
          enum: [new, handled]
        title:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        isTravelAdvisor:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
        acceptPrivacy:
          type: boolean
        receiveUpdates:
          type: boolean
        whatsappConsent:
          type: boolean
        phoneE164:
          type: ["string", "null"]
        tourTitle:
          type: ["string", "null"]
        tourLocation:
          type: ["string", "null"]
        tourDurationDays:
          type: ["number", "null"]
        attractionId:
          type: ["number", "null"]
        enquiryType:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        createdAt:
          type: string
        handledAt:
          type: ["string", "null"]
      required: [id, source, status, firstName, lastName, email, acceptPrivacy, receiveUpdates, whatsappConsent, createdAt]
    EnquiryInput:
      type: object
      properties:
        source:
          type: string
          enum: [tour, contact, attraction]
        title:
          type: string
          maxLength: 50
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
          maxLength: 254
        phone:
          type: ["string", "null"]
          maxLength: 50
        isTravelAdvisor:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
          maxLength: 5000
        acceptPrivacy:
          type: boolean
        receiveUpdates:
          type: boolean
        whatsappConsent:
          type: boolean
        tourTitle:
          type: string
          maxLength: 200
        tourLocation:
          type: string
          maxLength: 200
        tourDurationDays:
          type: number
          minimum: 1
          multipleOf: 1
        attractionId:
          type: number
          minimum: 1
          multipleOf: 1
        enquiryType:
          type: string
          enum: [tour-booking, custom-journey, membership, corporate, general]
        budget:
          type: ["string", "null"]
          maxLength: 100
      required: [source, firstName, lastName, email, acceptPrivacy, receiveUpdates]
    EnquiryStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [new, handled]
      required: [status]

    # ── Storage ───────────────────────────────────────────────────────
...