# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MNT Embark API — luxury tour company
 * OpenAPI spec version: 0.1.0
 */
import type { EnquirySource } from './enquirySource';
import type { EnquiryStatus } from './enquiryStatus';

export interface Enquiry {
  id: number;
  source: EnquirySource;
  status: EnquiryStatus;
  /** @nullable */
  title?: string | null;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  isTravelAdvisor?: boolean | null;
  /** @nullable */
  notes?: string | null;
  acceptPrivacy: boolean;
  receiveUpdates: boolean;
  whatsappConsent: boolean;
  /** @nullable */
  phoneE164?: string | null;
  /** @nullable */
  tourTitle?: string | null;
  /** @nullable */
  tourLocation?: string | null;
  /** @nullable */
  tourDurationDays?: number | null;
  /** @nullable */
  attractionId?: number | null;
  /** @nullable */
  enquiryType?: string | null;
  /** @nullable */
  budget?: string | null;
  createdAt: string;
  /** @nullable */
  handledAt?: string | null;
}