# lib/api-zod/src/generated/types/enquiryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MNT Embark API — luxury tour company
 * OpenAPI spec version: 0.1.0
 */
import type { EnquiryInputEnquiryType } from './enquiryInputEnquiryType';
import type { EnquiryInputSource } from './enquiryInputSource';

export interface EnquiryInput {
  source: EnquiryInputSource;
  /** @maxLength 50 */
  title?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  lastName: string;
  /**
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string;
  /**
     * @maxLength 50
     * @nullable
     */
  phone?: string | null;
  /** @nullable */
  isTravelAdvisor?: boolean | null;
  /**
     * @maxLength 5000
     * @nullable
     */
  notes?: string | null;
  acceptPrivacy: boolean;
  receiveUpdates: boolean;
  whatsappConsent?: boolean;
  /** @maxLength 200 */
  tourTitle?: string;
  /** @maxLength 200 */
  tourLocation?: string;
  /** @minimum 1 */
  tourDurationDays?: number;
  /** @minimum 1 */
  attractionId?: number;
  enquiryType?: EnquiryInputEnquiryType;
  /**
     * @maxLength 100
     * @nullable
     */
  budget?: string | null;
}