# lib/db/src/schema/enquiries.ts
import {
  boolean,
  index,
  integer,
  pgEnum,
  pgTable,
  serial,
  text,
  timestamp,
} from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { attractionsTable } from "./attractions";

export const enquirySourceEnum = pgEnum("enquiry_source", ["tour", "contact", "attraction"]);
export const enquiryStatusEnum = pgEnum("enquiry_status", ["new", "handled"]);

export const enquiriesTable = pgTable(
  "enquiries",
  {
    id: serial("id").primaryKey(),
    source: enquirySourceEnum("source").notNull(),
    status: enquiryStatusEnum("status").notNull().default("new"),
    title: text("title"),
    firstName: text("first_name").notNull(),
    lastName: text("last_name").notNull(),
    email: text("email").notNull(),
    phone: text("phone"),
    isTravelAdvisor: boolean("is_travel_advisor"),
    notes: text("notes"),
    acceptPrivacy: boolean("accept_privacy").notNull().default(false),

    /**
     * Permission to message this person on WhatsApp.
     *
     * Separate from `receiveUpdates` because it is a different promise: one is
     * marketing, this is a channel on their personal phone. Meta requires
     * demonstrable opt-in before a business messages anyone, and defaulting it
     * false means the absence of a decision is never read as consent.
     */
    whatsappConsent: boolean("whatsapp_consent").notNull().default(false),

    /**
     * The phone number in E.164 (+447700900123), derived from `phone` when the
     * enquiry is saved.
     *
     * Stored rather than normalised at send time so what we dial is fixed at
     * the moment they gave it, and so a number we could not make sense of is
     * visibly null instead of failing later inside the worker.
     */
    phoneE164: text("phone_e164"),
    receiveUpdates: boolean("receive_updates").notNull().default(false),
    tourTitle: text("tour_title"),
    tourLocation: text("tour_location"),
    tourDurationDays: integer("tour_duration_days"),
    /**
     * The attraction an "attraction" enquiry was sent from. Its name and place
     * are also copied into tour_title and tour_location at the time, so the
     * enquiry still reads correctly if the attraction is renamed or deleted.
     */
    attractionId: integer("attraction_id").references(() => attractionsTable.id, {
      onDelete: "set null",
    }),
    enquiryType: text("enquiry_type"),
    budget: text("budget"),
    createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
    handledAt: timestamp("handled_at", { withTimezone: true }),
  },
  (table) => [
    index("enquiries_status_created_idx").on(table.status, table.createdAt.desc()),
    index("enquiries_created_idx").on(table.createdAt.desc()),
  ],
);

export const insertEnquirySchema = createInsertSchema(enquiriesTable).omit({
  id: true,
  status: true,
  createdAt: true,
  handledAt: true,
});
export type InsertEnquiry = z.infer<typeof insertEnquirySchema>;
export type Enquiry = typeof enquiriesTable.$inferSelect;